backend_backup/requirements.txt
flask
flask_cors
flask_sqlalchemy
psycopg2-binary
requests
flask-migrate
beautifulsoup4
flask-jwt-extended
marshmallow
python-dotenv
Werkzeug